refactor mqtt_cfg_window out
